AddClothingForm and Closet components showing; set up clothing reducer
