+++ client/components/UserHome.js
@@ -4,21 +4,6 @@ import { Link } from 'react-router';
 import Clothing from './Clothing';
 import Closet from './Closet';
 
-const dummyPants = {
-  id: 1,
-  clothingType: 'pants',
-  color: ['green', 'multi'],
-  image: 'http://cdn.shopify.com/s/files/1/0286/5680/products/lunapants.png?v=1453829472',
-  tags: ['cute', 'floral', 'spring', 'casual']
-}
-
-const dummyShirt = {
-    id: 2,
-    clothingType: 'shirt',
-    color: ['purple'],
-    image: 'http://images.shirts.com/products/16493/14-14-43785/batman-neon-purple-t-shirt-back.jpg',
-    tags: ['casual', 'light']
-  };
 /**
  * UserHome will show:
  * 1. Outfit of the day
@@ -36,7 +21,8 @@ const UserHome = props => {
       <h3>Welcome, { user.email }</h3>
       <h2> 1. Outfit of the day will show here </h2>
       <h2> 2. Closet will show all pieces </h2>
-      <Closet allClothings={[dummyPants, dummyShirt]} />
+      <h1><Link to="addClothing">Add Clothing</Link></h1>
+      <h1><Link to="closet">Closet</Link></h1>
     </div>
   );
 }
