+++ client/reducer/index.js
@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux';
 import user from './user';
+import allClothings from './clothing.js';
 
 export default combineReducers({ user });
